requirements.txt
# Political Intelligence System - Complete Requirements
# Multi-competitor 2-stage ETL pipeline with queue system and web UI

# Core Dependencies
python-dotenv==1.0.0
PyYAML==6.0.1

# Database - PostgreSQL (Supabase)
asyncpg>=0.28.0
psycopg2-binary>=2.9.7
supabase>=2.4.0

# Queue System (Celery + Redis)
celery[redis]>=5.3.0
redis>=4.6.0
kombu>=5.3.0

# Web UI Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
jinja2>=3.1.2
python-multipart>=0.0.6
websockets>=11.0

# Method 1: Browser Automation
playwright>=1.49.0
selenium>=4.15.0

# Method 2: Essential Specialized Libraries
yt-dlp>=2023.12.30
snscrape>=0.7.0
instaloader>=4.10.3
facebook-scraper>=0.2.59

# Method 3: Web APIs and Services
aiohttp>=3.9.1
google-search-results>=2.4.0
apify-client>=1.7.0

# Method 4: Basic HTTP and Parsing
requests>=2.31.0
beautifulsoup4>=4.12.2
lxml>=4.9.3
feedparser>=6.0.11

# Content Extraction and Processing
newspaper3k>=0.2.8
trafilatura>=1.6.3
textstat>=0.7.3

# Text Processing and Similarity
python-levenshtein>=0.23.0
fuzzywuzzy>=0.18.0
nltk>=3.8.1

# Data Processing
pandas>=2.1.0
numpy>=1.24.0
pydantic>=2.4.0

# Essential Utilities
fake-useragent>=1.4.0
python-dateutil>=2.8.2
pytz>=2023.3
croniter>=1.4.1

# Performance and Monitoring
tqdm>=4.66.1
tenacity>=8.2.3
psutil>=5.9.0

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.9.0
isort>=5.12.0